@@ -7,95 +7,91 @@
         "dist",
         "src"
     ],
+    "exports": {
+        ".": {
+            "types": "./dist-esm/index.d.ts",
+            "require": "./dist/index.js",
+            "import": "./dist-esm/index.js",
+            "default": "./dist-esm/index.js"
+        }
+    },
     "engines": {
-        "node": ">=15"
+        "node": ">=20"
     },
     "scripts": {
-        "build:prod": "rollup -c --environment BUILD:production",
-        "build": "rollup -c",
+        "build:cjs": "tsc --build ./tsconfig.build.json",
+        "build:esm": "tsc --build ./tsconfig.build.esm.json",
+        "build": "yarn build:cjs && yarn build:esm",
         "docs:watch": "typedoc --watch",
         "docs": "typedoc",
         "lint": "eslint .",
-        "prepare": "husky install && yarn build:prod",
-        "start": "rollup -c --watch",
-        "test:watch": "jest --watch",
-        "test": "jest"
+        "test:watch": "NODE_OPTIONS=--openssl-legacy-provider jest --watch",
+        "test": "NODE_OPTIONS=--openssl-legacy-provider jest",
+        "test:win": "set NODE_OPTIONS=--openssl-legacy-provider && jest"
     },
     "lint-staged": {
         "**/*": "prettier --write --ignore-unknown"
     },
-    "name": "@avalabs/avalanche-wallet-sdk",
-    "author": "AVA Labs",
+    "name": "@runonflux/avalanche-wallet-sdk",
+    "author": "Ava Labs",
     "repository": {
         "type": "git",
-        "url": "https://github.com/ava-labs/avalanche-wallet-sdk"
+        "url": "https://github.com/ZelCore-io/avalanche-wallet-sdk"
     },
     "publishConfig": {
         "access": "public"
     },
     "devDependencies": {
-        "@babel/core": "7.17.8",
-        "@babel/preset-env": "7.16.11",
-        "@babel/preset-typescript": "7.16.7",
-        "@rollup/plugin-commonjs": "21.0.3",
-        "@rollup/plugin-json": "4.1.0",
-        "@rollup/plugin-typescript": "8.3.1",
+        "@babel/core": "~7.26.0",
+        "@babel/preset-env": "~7.26.0",
+        "@babel/preset-typescript": "~7.26.0",
         "@types/big.js": "^6.0.2",
-        "@types/bn.js": "5.1.0",
-        "@types/create-hash": "1.2.2",
-        "@types/crypto-js": "4.1.1",
-        "@types/hdkey": "2.0.1",
-        "@types/jest": "27.4.1",
-        "@types/randomstring": "1.1.8",
-        "@typescript-eslint/eslint-plugin": "5.17.0",
-        "@typescript-eslint/parser": "5.17.0",
-        "@zerollup/ts-transform-paths": "1.7.18",
-        "avalanche": "3.15.2",
-        "babel-jest": "27.5.1",
-        "babel-plugin-module-resolver": "4.1.0",
-        "eslint": "8.12.0",
-        "eslint-config-prettier": "8.5.0",
-        "husky": "6.0.0",
-        "jest": "27.5.1",
-        "lint-staged": "12.3.7",
-        "prettier": "2.6.2",
-        "rollup": "2.70.1",
-        "rollup-plugin-delete": "2.0.0",
-        "rollup-plugin-typescript2": "0.31.2",
-        "ts-node": "10.7.0",
-        "tslib": "2.3.1",
-        "ttypescript": "1.5.13",
-        "typedoc": "0.22.13",
-        "typescript": "4.6.3"
+        "@types/bn.js": "~5.1.0",
+        "@types/create-hash": "~1.2.2",
+        "@types/crypto-js": "~4.2.2",
+        "@types/jest": "~29.5.14",
+        "@types/randomstring": "~1.3.0",
+        "@types/tiny-secp256k1": "~2.0.1",
+        "@typescript-eslint/eslint-plugin": "^8.18.1",
+        "@typescript-eslint/parser": "^8.18.1",
+        "babel-jest": "~29.7.0",
+        "babel-plugin-module-resolver": "~5.0.2",
+        "eslint": "~9.17.0",
+        "eslint-config-prettier": "~9.1.0",
+        "eslint-import-resolver-typescript": "^3.7.0",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-jest": "^28.9.0",
+        "jest": "~29.7.0",
+        "lint-staged": "~15.2.11",
+        "prettier": "~3.4.2",
+        "ts-node": "~10.9.2",
+        "tslib": "~2.8.1",
+        "typedoc": "~0.27.5",
+        "typescript": "~4.8.4"
     },
     "dependencies": {
-        "@avalabs/hw-app-avalanche": "0.6.6",
-        "@ethereumjs/common": "2.6.0",
-        "@ethereumjs/tx": "3.4.0",
-        "@ledgerhq/hw-app-eth": "6.29.6",
-        "@ledgerhq/hw-transport": "6.27.5",
-        "@metamask/eth-sig-util": "4.0.0",
-        "@obsidiansystems/hw-app-avalanche": "0.2.2",
-        "@openzeppelin/contracts": "4.7.3",
+        "@avalabs/avalanchejs": "~3.17.0",
+        "@ethereumjs/common": "~4.4.0",
+        "@ethereumjs/tx": "~5.4.0",
+        "@metamask/eth-sig-util": "~8.1.1",
         "big.js": "^6.1.1",
-        "bip32": "2.0.6",
-        "bip32-path": "0.4.2",
-        "bip39": "3.0.4",
-        "bitcoinjs-lib": "5.2.0",
-        "bn.js": "5.2.1",
-        "create-hash": "1.2.0",
-        "crypto-js": "4.1.1",
+        "bip32": "~4.0.0",
+        "bip32-path": "~0.4.2",
+        "bip39": "~3.0.4",
+        "bitcoinjs-lib": "~5.2.0",
+        "bn.js": "~5.2.1",
+        "create-hash": "~1.2.0",
+        "crypto-js": "~4.2.0",
         "ethereumjs-util": "^7.0.7",
-        "ethers": "5.6.8",
-        "hdkey": "2.0.1",
-        "moment": "2.29.4",
-        "randomstring": "1.2.2",
-        "sockette": "2.0.6",
-        "typescript-transform-paths": "2.2.3",
-        "url": "0.11.0",
-        "web3": "1.8.1",
-        "web3-eth-contract": "1.8.1",
-        "xss": "1.0.11"
+        "ethers": "~5.7.2",
+        "moment": "~2.29.4",
+        "randomstring": "~1.3.0",
+        "sockette": "~2.0.6",
+        "tiny-secp256k1": "~2.2.3",
+        "url": "~0.11.0",
+        "web3": "~4.16.0",
+        "web3-eth-contract": "~4.7.2",
+        "xss": "~1.0.11"
     },
     "peerDependencies": {
         "avalanche": ">= 3.15.2"
