@@ -1,14 +1,14 @@
 module.exports = {
-    presets: [['@babel/preset-env', { targets: { node: 'current' } }], '@babel/preset-typescript'],
-    plugins: [
-        [
-            'module-resolver',
-            {
-                root: ['./src'],
-                alias: {
-                    '@': './src',
-                },
-            },
-        ],
-    ],
+  presets: [['@babel/preset-env', { targets: { node: 'current' } }], '@babel/preset-typescript'],
+  plugins: [
+      [
+          'module-resolver',
+          {
+              root: ['./src'],
+              alias: {
+                  '@': './src',
+              },
+          },
+      ],
+  ],
 };
