@@ -1,5 +1,5 @@
-import { Buffer } from 'avalanche';
-import { Erc20Token } from '@/Asset/Erc20Token';
+import { Buffer } from '@avalabs/avalanchejs';
+import { Erc20Token } from '../Asset/Erc20Token';
 
 export interface iAssetCache {
     [assetId: string]: iAssetDescriptionClean;
