@@ -1,11 +1,11 @@
-import { iHistoryImportExport } from '@/History/types';
-import { parseMemo } from '@/History/history_helpers';
-import { idToChainAlias } from '@/Network/helpers/aliasFromNetworkID';
-import { xChain } from '@/Network/network';
-import { bnToAvaxX, strip0x } from '@/utils';
-import { getOutputsOfChain, getOutputTotals, getOwnedOutputs } from '@/Explorer/ortelius/utxoUtils';
-import { findDestinationChain, findSourceChain, OrteliusAvalancheTx } from '@/Explorer';
-import { BN } from 'avalanche';
+import { iHistoryImportExport } from '../History/types';
+import { parseMemo } from '../History/history_helpers';
+import { idToChainAlias } from '../Network/helpers/aliasFromNetworkID';
+import { xChain } from '../Network/network';
+import { bnToAvaxX, strip0x } from '../utils';
+import { getOutputsOfChain, getOutputTotals, getOwnedOutputs } from '../Explorer/ortelius/utxoUtils';
+import { findDestinationChain, findSourceChain, OrteliusAvalancheTx } from '../Explorer';
+import { BN } from '@avalabs/avalanchejs';
 
 export function getImportSummary(tx: OrteliusAvalancheTx, addresses: string[], evmAddr: string): iHistoryImportExport {
     let sourceChain = findSourceChain(tx);
