@@ -1,9 +1,9 @@
-import { AVMWebSocketProvider } from '@/Network/providers/AVMWebSocketProvider';
-import { EVMWebSocketProvider } from '@/Network/providers/EVMWebSocketProvider';
-import { WalletType } from '@/Wallet/types';
-import { NetworkConfig } from '@/Network/types';
-import { wsUrlFromConfigEVM, wsUrlFromConfigX } from '@/helpers/network_helper';
-import { activeNetwork } from '@/Network/network';
+import { AVMWebSocketProvider } from '../../Network/providers/AVMWebSocketProvider';
+import { EVMWebSocketProvider } from '../../Network/providers/EVMWebSocketProvider';
+import { WalletType } from '../../Wallet/types';
+import { NetworkConfig } from '../../Network/types';
+import { wsUrlFromConfigEVM, wsUrlFromConfigX } from '../../helpers/network_helper';
+import { activeNetwork } from '../../Network/network';
 
 export class WebsocketProvider {
     avmProvider: AVMWebSocketProvider;
