@@ -1,12 +1,12 @@
-import { UniversalNodeAbstract } from '@/UniversalTx/UniversalNode';
-import { ExportChainsP } from '@/Wallet/types';
+import { UniversalNodeAbstract } from '../UniversalTx/UniversalNode';
+import { ExportChainsP } from '../Wallet/types';
 import {
     UniversalTxActionExportP,
     UniversalTxActionImportP,
     UniversalTxExportP,
     UniversalTxImportP,
-} from '@/UniversalTx/types';
-import { BN } from 'avalanche';
+} from '../UniversalTx/types';
+import { BN } from '@avalabs/avalanchejs';
 
 export default class UniversalNodeP extends UniversalNodeAbstract {
     constructor(balance: BN, feeExport: BN, feeImport: BN) {
