@@ -1,12 +1,12 @@
-import { UniversalNodeAbstract } from '@/UniversalTx/UniversalNode';
-import { ExportChainsX } from '@/Wallet/types';
+import { UniversalNodeAbstract } from '../UniversalTx/UniversalNode';
+import { ExportChainsX } from '../Wallet/types';
 import {
     UniversalTxActionExportX,
     UniversalTxActionImportX,
     UniversalTxExportX,
     UniversalTxImportX,
-} from '@/UniversalTx/types';
-import { BN } from 'avalanche';
+} from '../UniversalTx/types';
+import { BN } from '@avalabs/avalanchejs';
 
 export default class UniversalNodeX extends UniversalNodeAbstract {
     constructor(balance: BN, feeExport: BN, feeImport: BN) {
