@@ -1,4 +1,4 @@
-import { AVAX_TOKEN_PATH, ETH_ACCOUNT_PATH } from '@/Wallet/constants';
+import { AVAX_TOKEN_PATH, ETH_ACCOUNT_PATH } from '../../Wallet/constants';
 
 /**
  * Given an account number, returns the Avalanche account derivation path as a string
