@@ -1,8 +1,8 @@
-import { createGraphForC, createGraphForP, createGraphForX, UniversalTx } from '@/UniversalTx';
-import { BN } from 'avalanche';
-import { pChain, xChain } from '@/Network/network';
+import { createGraphForC, createGraphForP, createGraphForX, UniversalTx } from '../../src/UniversalTx';
+import { BN } from '@avalabs/avalanchejs';
+import { pChain, xChain } from '../../src/Network/network';
 
-jest.mock('@/Network/network', () => {
+jest.mock('../../src/Network/network', () => {
     return {
         web3: {
             utils: {
